Move IconX into the icon components directory
