+++ src/components/dialog.tsx
@@ -1,5 +1,6 @@
 import React, {FunctionComponent} from 'react'
 import {DialogOverlay, DialogContent} from '@reach/dialog'
+import IconX from 'components/icons/icon-x'
 
 type DialogProps = {
   children: React.ReactNode
@@ -43,20 +44,3 @@ const Dialog: FunctionComponent<DialogProps> = ({
 }
 
 export default Dialog
-
-const IconX: FunctionComponent<{className?: string}> = ({className = ''}) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    fill="none"
-    viewBox="0 0 24 24"
-    stroke="currentColor"
-    className={className}
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="M6 18L18 6M6 6l12 12"
-    />
-  </svg>
-)
