+++ src/components/icons/icon-x.tsx
@@ -1,0 +1,20 @@
+import React, {FunctionComponent} from 'react'
+
+const IconX: FunctionComponent<{className?: string}> = ({className = ''}) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+    viewBox="0 0 24 24"
+    stroke="currentColor"
+    className={className}
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M6 18L18 6M6 6l12 12"
+    />
+  </svg>
+)
+
+export default IconX
